Use `/etc/mw-collection-ocg.js` configuration file if present.
The documentation at the top of `defaults.js` mentions this file, but it
is nowhere loaded by default.  Since the provided upstart file doesn't
specify a `-c` option, we need to look for this override in a different way.

Note that this allows multiple levels of configuration -- the distribution's
default configuration file in `/etc/mw-collection-ocg.js` can specify another
more-local configuration in `/usr/local/etc/` which can then in turn fall
back to a user configuration in `$HOME`, etc.

Change-Id: Ie50b823d7dec510b30f5366c83e371c685576fee