Make sure the collection id is specific to the backend selected.
If the user first generated an output using (say) the rdf2text backend,
we'd return that output even if the user tried to generate a PDF for
the same article.  I believe this is an artifact of the original
design, where we planned to cache the bundler output separately.
But at the moment we only cache the final output, so the collection
ID needs to be specific to the backend chosen.

Change-Id: Ic6a55b9ac160a389a8a612fde87676752331590e