Fix redis gc.
The timestamp was being stored as a string in redis, but the comparison
was via integer/numeric comparison.  Make everything use the numeric
timestamp.

Add missing hdel wrapper to RedisWrapper to allow gc to delete keys
from Redis.

Don't crash if one of the temporary directories is set to null.

Change-Id: I9e198cdc46b791acb1fc3ba51fb557961006dc94