Write output files in subdirectories.
The output directory was growing to thousands of entries and signficantly
slowing down the health check.  Create subdirectories based on the first
few characters of the job id, and create/touch the subdirectories to
ensure that their cached sizes remain accurate.

Use the `force` parameter to the gc's `cleanDir` function to ensure
that we recurse into the top-level subdirectory.

Change-Id: I61451d51dcbfee959dfb066b663fec7a185df1a3