Limit memory usage during directory cleaning.
We were doing a breadth-first traversal of the directory tree.  That
resulted in excessive memory usage while cleaning up the postmortem
directory, which had a directory tree with a very large number of
leaves.

Switch to a depth-first traversal instead by limiting the number of
branches we traverse in parallel.

Change-Id: Id138b7f183ec7ac2fc5139c72fef6341785761ee