Separate out `download` and `download_file` commands.
The Collection RenderAPI expects the result of the `download` command
to be a JSON object.  We've been overriding this command as the
target of the JobStatus url field.  Use a new `download_file` command
for this purpose, so that we can eventually return a proper
"not supported" error to requests for `download`.

See comments on https://gerrit.wikimedia.org/r/178454 for some
discussion of the error paths involved.

Change-Id: I887b3c0051d336602c1bb0fd910c82da459fb085